requirements.txt
Flask
psycopg2-binary
openai
python-dotenv
tabulate
rich
sqlparse